Implementazione corretta soluzione stallo più riduzione del codice per una maggiore leggibilità

+++ Cena_dei_filosofi_chionne.c
@@ -15,6 +15,7 @@ void togli_forchette();
 void SigIntHandler(int iSignalCode);
 void filosofo(int i, sem_t *forchetta[]);
 
+
 int n_filosofi;
 sem_t *forchetta[3];
 char nome[20][10];
@@ -30,71 +31,80 @@ int f_stallo;
 int f_starv;
 int f_sol;
 
+
 void SigIntHandler(int iSignalCode) {
   //printf("Handler: ricevuto signal %d. %s\n.", iSignalCode,strsignal(iSignalCode));
   togli_forchette(); //rimuove le forchette (semafori)
   return;
 }
 
-
+//FUNZIONE COMPORTAMENTO FILOSOFO
 void filosofo(int i /*identificatore del filosofo*/, sem_t *forchetta[]) {
-  while (1) {
-    tempo.tv_sec = 0;   // secondi
-    tempo.tv_nsec = 8; // nanosecondi 
 
-    printf("Filosofo %d, ATTENDO la forchetta destra %d\n", i, i);
-    sem_wait(forchetta[i]);
-    printf("Filosofo %d, PRENDO la forchetta destra %d\n", i, i);
-    printf("Filosofo %d, ATTENDO la forchetta sinistra %d\n", i, (i+1)%n_filosofi);
-    sem_wait(forchetta[(i+1)%n_filosofi]);
-    printf("Filosofo %d, PRENDO la forchetta sinistra %d\n", i, (i+1)%n_filosofi);
+  int destra = i;
+  int sinistra = (i+1)%n_filosofi;
 
-    
-    printf("Filosofo %d: mangio...\n", i);
-    nanosleep(&tempo, NULL);
-
-    sem_post(forchetta[i]);
-    printf("Filosofo %d, RILASCIO la forchetta %d\n", i, i);
-    sem_post(forchetta[(i+1)%n_filosofi]);
-    printf("Filosofo %d, RILASCIO la forchetta %d\n", i, (i+1)%n_filosofi);    
+  //SOLUZIONE ALLO STALLO
+  if (f_sol != 0){
+    if (i == n_filosofi - 1){
+      // l'ultimo filosofo inverte l’ordine di presa delle forchette
+      destra = (i+1)%n_filosofi;
+      sinistra = i;
+    }
   }
-}
 
-void filosofo_con_stallo(int i /*identificatore del filosofo*/, sem_t *forchetta[]) {
   while (1) {
     tempo.tv_sec = 0;   // secondi
     tempo.tv_nsec = 8; // nanosecondi 
 
-    printf("Filosofo %d, ATTENDO la forchetta destra %d\n", i, i);
-    sem_wait(forchetta[i]);
-    printf("Filosofo %d, PRENDO la forchetta destra %d\n", i, i);
+    printf("Filosofo %d, ATTENDO la forchetta %d\n", i, destra);
+    sem_wait(forchetta[destra]);
+    printf("Filosofo %d, PRENDO la forchetta %d\n", i, destra);
     fflush(stdout);
 
-    sleep(3); //favorisce lo stallo
+    if (f_stallo != 0) sleep(1); //favorisce lo stallo
 
-    printf("Filosofo %d, ATTENDO la forchetta sinistra %d\n", i, (i+1)%n_filosofi);
+    printf("Filosofo %d, ATTENDO la forchetta %d\n", i, sinistra);
 
-    read(fd[0], &cont_stallo, sizeof(int));
-    cont_stallo++;
-    write(fd[1], &cont_stallo, sizeof(int));
-    if(cont_stallo == n_filosofi){
-        printf("STALLO RILEVATO\n");
+    if (f_stallo != 0){
+      read(fd[0], &cont_stallo, sizeof(int));
+      cont_stallo++;
+      write(fd[1], &cont_stallo, sizeof(int));
+      if(cont_stallo == n_filosofi){
+        fprintf(stderr,"STALLO RILEVATO\n");
         kill(0, SIGINT); //uccido i processi
+      }
     }
+    
+    sem_wait(forchetta[sinistra]);
+    printf("Filosofo %d, PRENDO la forchetta %d\n", i, sinistra);
+    read(fd[0], &cont_stallo, sizeof(int));
 
-    sem_wait(forchetta[(i+1)%n_filosofi]);
-    printf("Filosofo %d, PRENDO la forchetta sinistra %d\n", i, (i+1)%n_filosofi);
-
+    //prevenzione dello stallo -> se riesco a prendere la forchetta sinistra decremento il contatore 
+    if (f_stallo != 0){
+      cont_stallo--;
+      write(fd[1], &cont_stallo, sizeof(int));
+      printf("COUNTER %d\n", cont_stallo);
+    }
     
     printf("Filosofo %d: mangio...\n", i);
     nanosleep(&tempo, NULL);
 
-    sem_post(forchetta[i]);
-    printf("Filosofo %d, RILASCIO la forchetta %d\n", i, i);
-    sem_post(forchetta[(i+1)%n_filosofi]);
-    printf("Filosofo %d, RILASCIO la forchetta %d\n", i, (i+1)%n_filosofi);    
+    sem_post(forchetta[destra]);
+    printf("Filosofo %d, RILASCIO la forchetta %d\n", i, destra);
+    sem_post(forchetta[sinistra]);
+    printf("Filosofo %d, RILASCIO la forchetta %d\n", i, sinistra);    
   }
 }
+
+void togli_forchette(){
+    for (int i = 0; i < n_filosofi; i++){
+        sem_close(forchetta[i]);
+        sem_unlink(nome[i]);
+    }
+    return;
+}
+
 
 int main(int argc, char *argv[]) {
 
@@ -110,8 +120,10 @@ int main(int argc, char *argv[]) {
 
   n_filosofi = atoi(argv[1]);
 
+  
   if (n_filosofi < 2) {
-    printf("pochi filosofi a cena\n");
+    printf("Pochi filosofi a cena. Ce ne devono essere almeno 2\n");
+    return EXIT_FAILURE;
   }
 
   if (argv[2] != NULL)
@@ -127,7 +139,7 @@ int main(int argc, char *argv[]) {
     printf("Flag soluzione stallo attivato\n");
   if (f_starv != 0)
     printf("Flag starvation attivato\n");
-
+  
 
   //creazione delle forchette
   for (int i = 0; i < n_filosofi; i++) {
@@ -142,23 +154,19 @@ int main(int argc, char *argv[]) {
   
   pid_t filosofi[n_filosofi]; //gruppo di filosofi a cena
   
-  //CREAZIONE PIPE PER LA RILEVAZIONE DI STALLO 
-  cont_stallo = 0;
-  if (pipe(fd) == -1){
-    perror("pipe");
-    exit(EXIT_FAILURE);
+  //CREAZIONE PIPE PER LA RILEVAZIONE DI STALLO
+  if (f_stallo != 0){
+    cont_stallo = 0;
+    if (pipe(fd) == -1){
+      perror("pipe");
+      exit(EXIT_FAILURE);
+    }
+    write(fd[1], &cont_stallo, sizeof(int));
   }
-  write(fd[1], &cont_stallo, sizeof(int));
   
-
-  //CREAZIONE DEI FILOSOFI
-  if (f_sol != 0){ //nel caso in cui il flag è attivato
-    /*
-    APPROCCIO PER LA SOLUZIONE DELLO STALLO
-    Date n forchette, mi creo n-1 filosofi in modo tale che uno di essi riesce
-    ad avere entrambe le forchette, mangiare e lasciarle e così via per gli altri
-    */
-    for (int i = 0; i < n_filosofi-1; i++) {
+    
+  //creazione dei filosofi mediante processi figli
+  for (int i = 0; i < n_filosofi; i++) {
     filosofi[i] = fork();
 
     if (filosofi[i] == -1) {
@@ -168,51 +176,19 @@ int main(int argc, char *argv[]) {
     } else if (filosofi[i] == 0) {
       // child
       printf("Sono il filosofo numero: %d\n", i);
-      if (f_stallo != 0){
-        filosofo_con_stallo(i, forchetta);
-      }
       filosofo(i, forchetta);
       exit(0); //terminazione processo
-      }
-    } 
-  } else {
-      for (int i = 0; i < n_filosofi; i++) {
-      filosofi[i] = fork();
-
-      if (filosofi[i] == -1) {
-        perror("Errore in fork\n");
-        exit(EXIT_FAILURE);
-
-      } else if (filosofi[i] == 0) {
-        // child
-        printf("Sono il filosofo numero: %d\n", i);
-        if (f_stallo != 0){
-          filosofo_con_stallo(i, forchetta);
-        }
-        filosofo(i, forchetta);
-        exit(0); //terminazione processo
-      }
     }
   }
   
-    for (int i = 0; i < n_filosofi; i++){
-        waitpid(filosofi[i], NULL, 0);
-    }
+  //il padre sta in attesa dei filosofi
+  for (int i = 0; i < n_filosofi; i++){
+      waitpid(filosofi[i], NULL, 0);
+  }
     
-    //togli_forchette(); //metodo di rimozione delle forchette (semafori)
-    
-    //Messaggi signal (per evitare che venga chiamato esmplicitamente tante volte quanti sono i processi)
-    printf("Handler: ricevuto signal di Interrupt\n");
-    printf("Tolte le %d forchette dalla tavola\n", n_filosofi);
-    printf("Tutti i %d filosofi se ne sono andati\n", n_filosofi);
+  printf("Capo: Fine della cena\n");
+  printf("Tolte le %d forchette dalla tavola\n", n_filosofi);
+  printf("Tutti i %d filosofi se ne sono andati\n", n_filosofi);
 
-    return 0;   
-}
-
-void togli_forchette(){
-    for (int i = 0; i < n_filosofi; i++){
-        sem_close(forchetta[i]);
-        sem_unlink(nome[i]);
-    }
-    return;
+  return 0;   
 }
